Split tests by category
